Update setup instructions
The current instructions in both the readme and index.html had a few
inconsistencies, for example referencing files that don't exist.

In addition, the instructions were written with the goal of having a
self-contained platform and allowing folks to run the tests through
glitch. We use this repository as a practice assessment for students,
and the goals in using this repo as a practice assessment require a
slightly different approach.